fixed issues, extracted discovery and skill registry
