+++ tests/tools/test_registry.py
@@ -13,6 +13,7 @@ from core.tool_framework.base import BaseTool
 from core.tool_framework.registered_tool import REGISTERED_TOOL_ATTR, RegisteredTool
 from core.tool_framework.tool_decorator import tool
 from tools import registry as registry_module
+from tools import registry_discovery
 from tools.investigation_registry.actions import get_available_actions
 
 _V2_TOOL_CONTRACT_NAMES = frozenset(
@@ -49,7 +50,7 @@ def test_tool_decorator_registers_function_tool_with_inferred_schema() -> None:
     lookup_incident.__module__ = module.__name__
     module.lookup_incident = lookup_incident
 
-    tools = registry_module._collect_registered_tools_from_module(module)
+    tools = registry_discovery._collect_registered_tools_from_module(module)
 
     assert [tool_def.name for tool_def in tools] == ["lookup_incident"]
     registered = tools[0]
@@ -71,7 +72,7 @@ def test_tool_decorator_supports_minimal_single_file_function_tool() -> None:
     check_status.__module__ = module.__name__
     module.check_status = check_status
 
-    tools = registry_module._collect_registered_tools_from_module(module)
+    tools = registry_discovery._collect_registered_tools_from_module(module)
 
     assert [tool_def.name for tool_def in tools] == ["check_status"]
     registered = tools[0]
@@ -236,11 +237,11 @@ def test_auto_discovery_populates_investigation_and_chat_surfaces(
     module.get_incident_metadata = get_incident_metadata
 
     monkeypatch.setattr(
-        registry_module, "_iter_tool_module_names", lambda _pkg: ["fake_discovered_tool"]
+        registry_discovery, "_iter_tool_module_names", lambda _pkg: ["fake_discovered_tool"]
     )
     monkeypatch.setattr(registry_module, "_external_tool_packages", [])
     monkeypatch.setattr(registry_module, "_INTEGRATION_TOOL_PACKAGES", ())
-    monkeypatch.setattr(registry_module, "_import_tool_module", lambda _pkg, _name: module)
+    monkeypatch.setattr(registry_discovery, "_import_tool_module", lambda _pkg, _name: module)
 
     # Surface-scoped loads read the on-disk descriptor index; assert surface
     # assignment on the mocked full snapshot instead.
@@ -270,11 +271,11 @@ def test_action_surface_is_filtered_separately(monkeypatch: pytest.MonkeyPatch) 
     module.perform_action = perform_action
 
     monkeypatch.setattr(
-        registry_module, "_iter_tool_module_names", lambda _pkg: ["fake_action_tool"]
+        registry_discovery, "_iter_tool_module_names", lambda _pkg: ["fake_action_tool"]
     )
     monkeypatch.setattr(registry_module, "_external_tool_packages", [])
     monkeypatch.setattr(registry_module, "_INTEGRATION_TOOL_PACKAGES", ())
-    monkeypatch.setattr(registry_module, "_import_tool_module", lambda _pkg, _name: module)
+    monkeypatch.setattr(registry_discovery, "_import_tool_module", lambda _pkg, _name: module)
 
     # Surface-scoped loads read the on-disk descriptor index; assert surface
     # filtering on the mocked full snapshot instead.
@@ -372,7 +373,7 @@ def test_manifest_discovery_imports_nested_tool_modules(
     monkeypatch.setattr(registry_module, "_external_tool_packages", [])
     monkeypatch.setattr(registry_module, "_INTEGRATION_TOOL_PACKAGES", ())
     monkeypatch.setattr(
-        registry_module,
+        registry_discovery,
         "_iter_tool_module_names",
         lambda package: ["fake_provider"] if package is registry_module.tools_package else [],
     )
@@ -384,7 +385,7 @@ def test_manifest_discovery_imports_nested_tool_modules(
             return nested_module
         raise AssertionError(f"unexpected import: {package.__name__}.{name}")
 
-    monkeypatch.setattr(registry_module, "_import_tool_module", mock_import)
+    monkeypatch.setattr(registry_discovery, "_import_tool_module", mock_import)
 
     tools = registry_module.get_registered_tools()
 
@@ -417,7 +418,7 @@ def test_manifest_discovery_logs_nested_import_failure_with_full_module_path(
     monkeypatch.setattr(registry_module, "_external_tool_packages", [])
     monkeypatch.setattr(registry_module, "_INTEGRATION_TOOL_PACKAGES", ())
     monkeypatch.setattr(
-        registry_module,
+        registry_discovery,
         "_iter_tool_module_names",
         lambda package: ["fake_provider"] if package is registry_module.tools_package else [],
     )
@@ -431,7 +432,7 @@ def test_manifest_discovery_logs_nested_import_failure_with_full_module_path(
             return valid_module
         raise AssertionError(f"unexpected import: {package.__name__}.{name}")
 
-    monkeypatch.setattr(registry_module, "_import_tool_module", mock_import)
+    monkeypatch.setattr(registry_discovery, "_import_tool_module", mock_import)
 
     with caplog.at_level(logging.WARNING, logger="tools.registry"):
         tools = registry_module.get_registered_tools()
@@ -477,12 +478,12 @@ def test_manifest_discovery_preserves_duplicate_name_first_wins(
     monkeypatch.setattr(registry_module, "_external_tool_packages", [])
     monkeypatch.setattr(registry_module, "_INTEGRATION_TOOL_PACKAGES", ())
     monkeypatch.setattr(
-        registry_module,
+        registry_discovery,
         "_iter_tool_module_names",
         lambda package: ["fake_provider"] if package is registry_module.tools_package else [],
     )
     monkeypatch.setattr(
-        registry_module,
+        registry_discovery,
         "_import_tool_module",
         lambda package, name: (
             provider_pkg
@@ -522,7 +523,7 @@ def test_top_level_discovery_unchanged_without_manifest(
     monkeypatch.setattr(registry_module, "_external_tool_packages", [])
     monkeypatch.setattr(registry_module, "_INTEGRATION_TOOL_PACKAGES", ())
     monkeypatch.setattr(
-        registry_module,
+        registry_discovery,
         "_iter_tool_module_names",
         lambda package: ["fake_top_level_tool"] if package is registry_module.tools_package else [],
     )
@@ -531,7 +532,7 @@ def test_top_level_discovery_unchanged_without_manifest(
         import_calls.append((package.__name__, name))
         return top_level_module
 
-    monkeypatch.setattr(registry_module, "_import_tool_module", mock_import)
+    monkeypatch.setattr(registry_discovery, "_import_tool_module", mock_import)
 
     tools = registry_module.get_registered_tools()
 
@@ -557,9 +558,9 @@ def test_resolve_tool_display_name_prefers_registered_metadata(
     module.get_incident_metadata = get_incident_metadata
 
     monkeypatch.setattr(
-        registry_module, "_iter_tool_module_names", lambda _pkg: ["fake_display_name_tool"]
+        registry_discovery, "_iter_tool_module_names", lambda _pkg: ["fake_display_name_tool"]
     )
-    monkeypatch.setattr(registry_module, "_import_tool_module", lambda _pkg, _name: module)
+    monkeypatch.setattr(registry_discovery, "_import_tool_module", lambda _pkg, _name: module)
 
     assert registry_module.resolve_tool_display_name("get_incident_metadata") == "Incident metadata"
 
@@ -630,12 +631,12 @@ def test_registry_regression_duplicate_tool_names_across_modules(
     module2.shared_tool_second = second_tool
 
     monkeypatch.setattr(
-        registry_module,
+        registry_discovery,
         "_iter_tool_module_names",
         lambda _pkg: ["first_module", "second_module"],
     )
     monkeypatch.setattr(
-        registry_module,
+        registry_discovery,
         "_import_tool_module",
         lambda _pkg, name: module1 if name == "first_module" else module2,
     )
@@ -679,12 +680,12 @@ def test_registry_regression_import_failures(
         return module
 
     monkeypatch.setattr(
-        registry_module,
+        registry_discovery,
         "_iter_tool_module_names",
         lambda _pkg: ["broken_module", "valid_tool"],
     )
     monkeypatch.setattr(
-        registry_module,
+        registry_discovery,
         "_import_tool_module",
         mock_import,
     )
@@ -969,8 +970,8 @@ def test_registry_canonical_tool_wins_when_external_package_redefines_name(
     fake_pkg.__path__ = []  # type: ignore[attr-defined]
 
     # Inject the fake by patching the walk + import.
-    real_iter = registry_module._iter_tool_module_names
-    real_import = registry_module._import_tool_module
+    real_iter = registry_discovery._iter_tool_module_names
+    real_import = registry_discovery._import_tool_module
 
     def fake_iter(package: ModuleType) -> list[str]:
         if package is fake_pkg:
@@ -982,8 +983,8 @@ def test_registry_canonical_tool_wins_when_external_package_redefines_name(
             return fake_module
         return real_import(package, name)
 
-    monkeypatch.setattr(registry_module, "_iter_tool_module_names", fake_iter)
-    monkeypatch.setattr(registry_module, "_import_tool_module", fake_import)
+    monkeypatch.setattr(registry_discovery, "_iter_tool_module_names", fake_iter)
+    monkeypatch.setattr(registry_discovery, "_import_tool_module", fake_import)
 
     saved = list(registry_module._external_tool_packages)
     try:
